Implement functionality to list planets with moons and calculate average moon temperatures - Added temperature property to MoonDto to capture moon temperature data. - Updated Planet class to calculate average temperature of its moons. - Added Memory Cache to To prevent from constant call

+++ Domain/DataTransferObjects/PlanetDto.cs
@@ -9,7 +9,7 @@ namespace TenForce_Nikunj_DeveloperExercise.Domain.DataTransferObjects
     public class PlanetDto
     {
         public string Id { get; set; }
-        public float SemiMajorAxis { get; set; }
+        public long SemiMajorAxis { get; set; }
         public ICollection<MoonDto> Moons { get; set; }
     }
 }
